Move background updates to `StateBackgroundUpdateStore`
So we can access `_get_state_groups_from_groups_txn(...)`
